requirements.txt
aiogram==2.9.0
aiohttp
aiohttp-cors
aiosqlite
asyncpg
Pyrogram
PySocks
pytest
pytz
tortoise-orm
python-dotenv
